Performance improvements - add types cache to MetaModel (improve ExistClass perf) - ByteArrayConverter - improve copying of binary arrays - NonNativeObjectWriter - improve copying of binary arrays - MultiBuffer - improve cleaning of bytes array perf - remove static helper fields from MultiBufferedFileIO

+++ src/Odb/Core/Layers/Layer3/IO/IOdbStream.cs
@@ -17,7 +17,7 @@ namespace NDatabase.Odb.Core.Layers.Layer3.IO
 
         void Write(byte[] buffer, int size);
 
-        long Read(byte[] buffer, int size);
+        int Read(byte[] buffer, int size);
 
         int Read();
